refactor(desktop): move commands into their own module
