+++ harper-desktop/src-tauri/src/commands.rs
@@ -1,0 +1,342 @@
+//! This module centralizes all the Rust commands to be made available to the JavaScript
+//! environment. Use [`application_message_handler`] to load them into the Tauri runtime.
+
+use crate::config::Config;
+use crate::highlighter_service::HighlighterService;
+use crate::os_broker::{AccessibilityPermissionStatus, OsBroker};
+use crate::{IntegrationView, accessibility_allows_highlighter_start, platform_broker};
+use harper_core::{
+    Dialect, DictWordMetadata, IgnoredLints,
+    linting::FlatConfig,
+    spell::{Dictionary, MutableDictionary},
+};
+use std::sync::Arc;
+use tauri::ipc::Invoke;
+use tauri::{Runtime, State};
+use tokio::sync::Mutex;
+
+pub fn application_message_handler<R: Runtime>() -> impl Fn(Invoke<R>) -> bool {
+    tauri::generate_handler![
+        get_lint_config,
+        get_dialect,
+        get_debounce_ms,
+        set_debounce_ms,
+        get_auto_update,
+        set_auto_update,
+        get_last_update_check,
+        set_last_update_check,
+        set_dialect,
+        set_lint_config,
+        get_dictionary,
+        set_dictionary,
+        ignore_lint,
+        add_to_dictionary,
+        get_integrations,
+        add_integration,
+        remove_integration,
+        set_integration_enabled,
+        get_accessibility_permission_status,
+        request_accessibility_permission,
+        start_highlighter_service,
+        stop_highlighter_service,
+        launch_app,
+    ]
+}
+
+#[tauri::command]
+async fn get_lint_config(config: State<'_, Arc<Mutex<Config>>>) -> Result<FlatConfig, String> {
+    let mut lint_config = config.lock().await.lint_config.clone();
+    lint_config.fill_with_curated();
+
+    Ok(lint_config)
+}
+
+#[tauri::command]
+async fn get_dialect(config: State<'_, Arc<Mutex<Config>>>) -> Result<Dialect, String> {
+    Ok(config.lock().await.dialect)
+}
+
+#[tauri::command]
+async fn get_debounce_ms(config: State<'_, Arc<Mutex<Config>>>) -> Result<u64, String> {
+    Ok(config.lock().await.debounce_ms)
+}
+
+#[tauri::command]
+async fn set_debounce_ms(
+    debounce_ms: u64,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut config = config.lock().await;
+    config.debounce_ms = debounce_ms;
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn get_auto_update(config: State<'_, Arc<Mutex<Config>>>) -> Result<bool, String> {
+    Ok(config.lock().await.auto_update)
+}
+
+#[tauri::command]
+async fn set_auto_update(
+    auto_update: bool,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut config = config.lock().await;
+    config.auto_update = auto_update;
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn get_last_update_check(
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<Option<u64>, String> {
+    Ok(config.lock().await.last_update_check)
+}
+
+#[tauri::command]
+async fn set_last_update_check(
+    last_update_check: Option<u64>,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut config = config.lock().await;
+    config.last_update_check = last_update_check;
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn set_dialect(
+    dialect: Dialect,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut config = config.lock().await;
+    config.dialect = dialect;
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn set_lint_config(
+    lint_config: FlatConfig,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut lint_config = lint_config;
+    lint_config.fill_with_curated();
+
+    let mut config = config.lock().await;
+    config.lint_config = lint_config;
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn get_dictionary(config: State<'_, Arc<Mutex<Config>>>) -> Result<Vec<String>, String> {
+    let mut words = config
+        .lock()
+        .await
+        .mutable_dictionary
+        .words_iter()
+        .map(|word| word.iter().collect::<String>())
+        .collect::<Vec<_>>();
+    words.sort();
+
+    Ok(words)
+}
+
+#[tauri::command]
+async fn set_dictionary(
+    words: Vec<String>,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut dictionary = MutableDictionary::new();
+    dictionary.extend_words(words.into_iter().map(|word| {
+        (
+            word.chars().collect::<Vec<_>>(),
+            DictWordMetadata::default(),
+        )
+    }));
+
+    let mut config = config.lock().await;
+    config.mutable_dictionary = dictionary;
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn ignore_lint(
+    ignored_lints: String,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let ignored_lints =
+        serde_json::from_str::<IgnoredLints>(&ignored_lints).map_err(|error| error.to_string())?;
+
+    let mut config = config.lock().await;
+    config.ignored_lints.append(ignored_lints);
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn add_to_dictionary(
+    word: String,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut config = config.lock().await;
+    config
+        .mutable_dictionary
+        .append_word_str(&word, DictWordMetadata::default());
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn get_integrations(
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<Vec<IntegrationView>, String> {
+    let integrations = config.lock().await.integrations.clone();
+    let broker = platform_broker();
+
+    Ok(integrations
+        .into_iter()
+        .map(|integration| IntegrationView {
+            display_name: broker.integration_display_name(&integration.bundle_id),
+            bundle_id: integration.bundle_id,
+            enabled: integration.enabled,
+        })
+        .collect())
+}
+
+#[tauri::command]
+async fn add_integration(
+    bundle_id: String,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut config = config.lock().await;
+    config.add_integration(bundle_id);
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn remove_integration(
+    bundle_id: String,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut config = config.lock().await;
+    config.remove_integration(&bundle_id);
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+async fn set_integration_enabled(
+    bundle_id: String,
+    enabled: bool,
+    config: State<'_, Arc<Mutex<Config>>>,
+) -> Result<(), String> {
+    let mut config = config.lock().await;
+    config.set_integration_enabled(&bundle_id, enabled);
+    config
+        .save_to_system()
+        .await
+        .map_err(|error| error.to_string())?;
+
+    Ok(())
+}
+
+#[tauri::command]
+fn get_accessibility_permission_status() -> AccessibilityPermissionStatus {
+    platform_broker().accessibility_permission_status()
+}
+
+#[tauri::command]
+fn request_accessibility_permission() -> AccessibilityPermissionStatus {
+    platform_broker().request_accessibility_permission()
+}
+
+#[tauri::command]
+pub(crate) async fn start_highlighter_service(
+    config: State<'_, Arc<Mutex<Config>>>,
+    highlighter_service: State<'_, HighlighterService>,
+) -> Result<bool, String> {
+    {
+        let mut config = config.lock().await;
+        config.highlighter_service_enabled = true;
+        config
+            .save_to_system()
+            .await
+            .map_err(|error| error.to_string())?;
+    }
+
+    if accessibility_allows_highlighter_start() {
+        highlighter_service
+            .start()
+            .map_err(|error| error.to_string())?;
+    }
+
+    Ok(highlighter_service.is_running())
+}
+
+#[tauri::command]
+pub(crate) async fn stop_highlighter_service(
+    config: State<'_, Arc<Mutex<Config>>>,
+    highlighter_service: State<'_, HighlighterService>,
+) -> Result<bool, String> {
+    {
+        let mut config = config.lock().await;
+        config.highlighter_service_enabled = false;
+        config
+            .save_to_system()
+            .await
+            .map_err(|error| error.to_string())?;
+    }
+
+    Ok(highlighter_service.stop())
+}
+
+#[tauri::command]
+fn launch_app(bundle_id: String) -> Result<(), String> {
+    platform_broker().launch_app_bundle(&bundle_id)
+}
