+++ harper-desktop/src-tauri/src/lib.rs
@@ -6,13 +6,13 @@ use crate::debounce::{DebounceState, DebounceStatus};
 use clap::{Parser, Subcommand};
 use harper_core::{
     Dialect, DictWordMetadata, Document, IgnoredLints,
-    linting::{FlatConfig, Lint, LintGroup},
-    spell::{Dictionary, MutableDictionary},
+    linting::{Lint, LintGroup},
+    spell::MutableDictionary,
 };
 use serde::Serialize;
 use std::{cell::RefCell, rc::Rc, sync::Arc, time::Duration};
 use tauri::{
-    Manager, State, WebviewUrl, WebviewWindowBuilder, WindowEvent,
+    Manager, WebviewUrl, WebviewWindowBuilder, WindowEvent,
     image::Image,
     menu::{HELP_SUBMENU_ID, Menu, MenuItem, PredefinedMenuItem},
     tray::{MouseButton, TrayIconBuilder, TrayIconEvent},
@@ -27,6 +27,7 @@ use tokio::{
 };
 
 pub mod color;
+mod commands;
 pub mod communication;
 pub mod config;
 mod debounce;
@@ -256,304 +257,6 @@ fn start_highlighter_service_if_enabled_and_permitted(
     }
 }
 
-#[tauri::command]
-async fn get_lint_config(config: State<'_, Arc<Mutex<Config>>>) -> Result<FlatConfig, String> {
-    let mut lint_config = config.lock().await.lint_config.clone();
-    lint_config.fill_with_curated();
-
-    Ok(lint_config)
-}
-
-#[tauri::command]
-async fn get_dialect(config: State<'_, Arc<Mutex<Config>>>) -> Result<Dialect, String> {
-    Ok(config.lock().await.dialect)
-}
-
-#[tauri::command]
-async fn get_debounce_ms(config: State<'_, Arc<Mutex<Config>>>) -> Result<u64, String> {
-    Ok(config.lock().await.debounce_ms)
-}
-
-#[tauri::command]
-async fn set_debounce_ms(
-    debounce_ms: u64,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut config = config.lock().await;
-    config.debounce_ms = debounce_ms;
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn get_auto_update(config: State<'_, Arc<Mutex<Config>>>) -> Result<bool, String> {
-    Ok(config.lock().await.auto_update)
-}
-
-#[tauri::command]
-async fn set_auto_update(
-    auto_update: bool,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut config = config.lock().await;
-    config.auto_update = auto_update;
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn get_last_update_check(
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<Option<u64>, String> {
-    Ok(config.lock().await.last_update_check)
-}
-
-#[tauri::command]
-async fn set_last_update_check(
-    last_update_check: Option<u64>,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut config = config.lock().await;
-    config.last_update_check = last_update_check;
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn set_dialect(
-    dialect: Dialect,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut config = config.lock().await;
-    config.dialect = dialect;
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn set_lint_config(
-    lint_config: FlatConfig,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut lint_config = lint_config;
-    lint_config.fill_with_curated();
-
-    let mut config = config.lock().await;
-    config.lint_config = lint_config;
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn get_dictionary(config: State<'_, Arc<Mutex<Config>>>) -> Result<Vec<String>, String> {
-    let mut words = config
-        .lock()
-        .await
-        .mutable_dictionary
-        .words_iter()
-        .map(|word| word.iter().collect::<String>())
-        .collect::<Vec<_>>();
-    words.sort();
-
-    Ok(words)
-}
-
-#[tauri::command]
-async fn set_dictionary(
-    words: Vec<String>,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut dictionary = MutableDictionary::new();
-    dictionary.extend_words(words.into_iter().map(|word| {
-        (
-            word.chars().collect::<Vec<_>>(),
-            DictWordMetadata::default(),
-        )
-    }));
-
-    let mut config = config.lock().await;
-    config.mutable_dictionary = dictionary;
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn ignore_lint(
-    ignored_lints: String,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let ignored_lints =
-        serde_json::from_str::<IgnoredLints>(&ignored_lints).map_err(|error| error.to_string())?;
-
-    let mut config = config.lock().await;
-    config.ignored_lints.append(ignored_lints);
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn add_to_dictionary(
-    word: String,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut config = config.lock().await;
-    config
-        .mutable_dictionary
-        .append_word_str(&word, DictWordMetadata::default());
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn get_integrations(
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<Vec<IntegrationView>, String> {
-    let integrations = config.lock().await.integrations.clone();
-    let broker = platform_broker();
-
-    Ok(integrations
-        .into_iter()
-        .map(|integration| IntegrationView {
-            display_name: broker.integration_display_name(&integration.bundle_id),
-            bundle_id: integration.bundle_id,
-            enabled: integration.enabled,
-        })
-        .collect())
-}
-
-#[tauri::command]
-async fn add_integration(
-    bundle_id: String,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut config = config.lock().await;
-    config.add_integration(bundle_id);
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn remove_integration(
-    bundle_id: String,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut config = config.lock().await;
-    config.remove_integration(&bundle_id);
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn set_integration_enabled(
-    bundle_id: String,
-    enabled: bool,
-    config: State<'_, Arc<Mutex<Config>>>,
-) -> Result<(), String> {
-    let mut config = config.lock().await;
-    config.set_integration_enabled(&bundle_id, enabled);
-    config
-        .save_to_system()
-        .await
-        .map_err(|error| error.to_string())?;
-
-    Ok(())
-}
-
-#[tauri::command]
-fn get_accessibility_permission_status() -> AccessibilityPermissionStatus {
-    platform_broker().accessibility_permission_status()
-}
-
-#[tauri::command]
-fn request_accessibility_permission() -> AccessibilityPermissionStatus {
-    platform_broker().request_accessibility_permission()
-}
-
-#[tauri::command]
-async fn start_highlighter_service(
-    config: State<'_, Arc<Mutex<Config>>>,
-    highlighter_service: State<'_, HighlighterService>,
-) -> Result<bool, String> {
-    {
-        let mut config = config.lock().await;
-        config.highlighter_service_enabled = true;
-        config
-            .save_to_system()
-            .await
-            .map_err(|error| error.to_string())?;
-    }
-
-    if accessibility_allows_highlighter_start() {
-        highlighter_service
-            .start()
-            .map_err(|error| error.to_string())?;
-    }
-
-    Ok(highlighter_service.is_running())
-}
-
-#[tauri::command]
-async fn stop_highlighter_service(
-    config: State<'_, Arc<Mutex<Config>>>,
-    highlighter_service: State<'_, HighlighterService>,
-) -> Result<bool, String> {
-    {
-        let mut config = config.lock().await;
-        config.highlighter_service_enabled = false;
-        config
-            .save_to_system()
-            .await
-            .map_err(|error| error.to_string())?;
-    }
-
-    Ok(highlighter_service.stop())
-}
-
-#[tauri::command]
-fn launch_app(bundle_id: String) -> Result<(), String> {
-    platform_broker().launch_app_bundle(&bundle_id)
-}
-
 #[cfg(target_os = "macos")]
 fn platform_broker() -> impl OsBroker {
     mac_broker::MacBroker::default()
@@ -621,31 +324,7 @@ pub fn run_tauri() {
         ))
         .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_os::init())
-        .invoke_handler(tauri::generate_handler![
-            get_lint_config,
-            get_dialect,
-            get_debounce_ms,
-            set_debounce_ms,
-            get_auto_update,
-            set_auto_update,
-            get_last_update_check,
-            set_last_update_check,
-            set_dialect,
-            set_lint_config,
-            get_dictionary,
-            set_dictionary,
-            ignore_lint,
-            add_to_dictionary,
-            get_integrations,
-            add_integration,
-            remove_integration,
-            set_integration_enabled,
-            get_accessibility_permission_status,
-            request_accessibility_permission,
-            start_highlighter_service,
-            stop_highlighter_service,
-            launch_app,
-        ])
+        .invoke_handler(commands::application_message_handler())
         .menu(desktop_app_menu)
         .on_menu_event(|app, event| {
             if event.id().as_ref() == REPORT_ISSUE_MENU_ID {
@@ -681,15 +360,17 @@ pub fn run_tauri() {
                         let highlighter_service = app.state::<HighlighterService>();
 
                         let toggle_result = if highlighter_service.is_running() {
-                            tauri::async_runtime::block_on(stop_highlighter_service(
+                            tauri::async_runtime::block_on(commands::stop_highlighter_service(
                                 app.state::<Arc<Mutex<Config>>>(),
                                 highlighter_service,
                             ))
                         } else {
-                            let result = tauri::async_runtime::block_on(start_highlighter_service(
-                                app.state::<Arc<Mutex<Config>>>(),
-                                highlighter_service,
-                            ));
+                            let result = tauri::async_runtime::block_on(
+                                commands::start_highlighter_service(
+                                    app.state::<Arc<Mutex<Config>>>(),
+                                    highlighter_service,
+                                ),
+                            );
 
                             if matches!(result, Ok(false))
                                 && !accessibility_allows_highlighter_start()
